requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Django==4.1.8
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
docopt==0.6.2
idna==3.4
pipreqs==0.4.13
pycparser==2.21
PyJWT==2.6.0
pymongo==3.12.1
pytz==2023.3
requests==2.29.0
sqlparse==0.2.4
tzdata==2023.3
urllib3==1.26.15
yarg==0.1.9